Changing Internet Check URL and status code
Earlier the check was on https://www.google.com. But customers usually block access to google.com. So changing the check to cxd.cisco.com/home, which is the actual URL used for upload.  cxd.cisco.com will give a 401 Response code, but it proves that TCP Port 443 to cxd.cisco.com is open.